+++ src/components/Sidebar/SidebarContext.tsx
@@ -3,12 +3,16 @@ import React, {
 	PropsWithChildren,
 	SetStateAction,
 	createContext,
+	useCallback,
 	useContext,
 	useState
 } from "react";
 
+import { Setter, getValueFromSetter } from "../../utilities";
+
 type SidebarContextProps = {
 	isOpen: boolean;
+	isClosing: boolean;
 	setIsOpen: Dispatch<SetStateAction<boolean>>;
 };
 
@@ -18,10 +22,35 @@ export function SidebarProvider({
 	children,
 	initialOpen = false
 }: PropsWithChildren<{ initialOpen?: boolean }>) {
-	const [isOpen, setIsOpen] = useState(initialOpen);
+	const [isOpen, _setIsOpen] = useState(initialOpen);
+	const [isClosing, setIsClosing] = useState(false);
+
+	/**
+	 * wrapper around `_setIsOpen`, to control `isClosing`.
+	 * we should instead have a single "status" variable, but refactoring might suck..
+	 *
+	 * the purpose of `isClosing` is single:
+	 * when sidebar is opened, and user hovers over it and clicks the button to close it,
+	 * then the sidebar gets automatically closed, instead of waiting for the user
+	 * to hover away from it before closing.
+	 */
+	const setIsOpen = useCallback((newValue: Setter<boolean>): void => {
+		_setIsOpen((currIsOpen) => {
+			const value: boolean = getValueFromSetter(newValue, currIsOpen);
+
+			if (value === false) {
+				setIsClosing(true);
+				setTimeout(() => setIsClosing(false), 300);
+			}
+
+			return value;
+		});
+	}, []);
 
 	return (
-		<SidebarContext.Provider value={{ isOpen, setIsOpen }}>{children}</SidebarContext.Provider>
+		<SidebarContext.Provider value={{ isOpen, setIsOpen, isClosing }}>
+			{children}
+		</SidebarContext.Provider>
 	);
 }