+++ src/components/client.ts
@@ -17,3 +17,4 @@ export * from "./AlertDialog";
 export * from "./ScrollArea";
 export * from "./HoverCard";
 export * from "./RadioGroup";
+export * from "./Command";